cmd/sleep: Consume Ctrl-C when exiting early
When sleep aborts early due to Ctrl-C input from the user, that state
is transferred to the exit status of the program and should therefore
be cleared from the global state.

Before this change, the lingering global state would lead to very
unintuitive execution flows in scripts due to the U-Boot specific
modification to the hush interpreter which checks for Ctrl-C in all
loop constructs.

This script, for example...

   if sleep 3; then
       echo "Had a good night's sleep"
   else
       echo "That was a rude awakening! I refuse to go on..."
       while true; do
           true
       done

       echo "...or maybe I will"
   fi

...would just skip over the infinite loop (indeed, all following
loops) when prematurely interrupting the sleep.

(cherry picked from commit ed733077b87b1f3da6ac956840c927406c268a3d)