hush: Remove Ctrl-C detection in loops
Assume that the original intent was to emulate SIGINT to a shell. This
only works as expected if the loop in question is the ouermost, and
last, statement. In all other cases, it completely breaks the expected
execution flow. It more or less resurrects Visual Basic's "On Error
Resume Next".

Disable this behavior and delegate the problem of loop termination to
the writer of the script instead.

Signed-off-by: Tobias Waldekranz <tobias@waldekranz.com>
(cherry picked from commit cde9fed496833b88094ed6e9fea11969a1be8e25)